requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Exchange API
krakenex>=2.2.0

# Telegram bot
python-telegram-bot>=20.7

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development tools
pre-commit>=3.6.0
ruff>=0.1.9
mypy>=1.7.1
types-requests>=2.31.0
types-python-dateutil>=2.8.19
types-PyYAML>=6.0.12
